Broke applib up into separate modules
