+++ src-app/util/applib/filesystem.ts
@@ -1,0 +1,5 @@
+import * as path from 'path';
+import * as findRoot from 'find-root';
+
+/** The filesystem root of the whole project */
+export const rootPath = findRoot(path.dirname(require.main.filename));
